+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [5.0.0] - 2025-04-24
+## [5.0.0] - 2025-04-25
 ### Changed
  - Upgrade to JRuby 10.0.0.0 (Ruby 3.4.2)
  - Upgrade to Groovy 4.0.26